Use size_t rather than int to track dynamic buffer size.
This lets use create dynamic buffers larger than 2GB.

+++ src/calendar.c
@@ -1336,7 +1336,7 @@ static void WriteCalTrailer(void)
 /***************************************************************/
 static int DoCalRem(ParsePtr p, int col)
 {
-    int oldLen;
+    size_t oldLen;
     Trigger trig;
     TimeTrig tim;
     Value v;
